feat(wip) add FileProcessor to handle file processing
I'm not 100% sure on the class structure yet, as I'm
experimenting with things.

One thing that should be noted is that the destination
where the files are saved will be something like Azure
Blob Storage on S3 Bucket.
The current implementation is not matching the above
statement, but it is just a placeholder information to
be reworked in the future.