revert move wwwroot to the root folder again
With the wwwroot folder inside Presentation/, it works on my machine, but when I
deploy the project, the static files aren't being found.

I didn't know that the wwwroot folder, when the pogram compiles, is placed in the
same directory as the binaries. For some reason, when the wwwroot is placed inside
the Presentation/ folder, the static files don't make it. They're just no there (inside the
wwwroot folder) in the prod env.

What I don't completely understand yet is why this happens. Moving the wwwroot to
the root folder solves the problem. I'm still investigating to try to understand this behavior,
but for now, I'll just revert this change.