refact! change project structure to a VSA approach
Even though I moved to VSA, I'm still layering inside the features,
for example, because I like the division of technical concerns.

Each view/ folder inside each feature has a _ViewStart.cshtml file
because of the way mvc looks for the _ViewStart.cshtml file. If I had
placed it inside the Components/Presentation/Views folder, it would
not work, because mvc wouldn't be able to find it.

Finally, the reason why I'm moving to this approach is because it
make
The way it does is that it checks in the parent directory and keeps
going up until it find the _ViewStart.cshtml file. If it does not find,
then it will just render the View without calling a _ViewStart.cshtml.

Finally, the reason why I'm moving to this approach is because it
is more easy to reason about and easier to give maintanence (theoretically),
as the files are divided by their concerns, and not by technical stuff